Split logrus config out
- Keep main.go as simple as possiable

- Add `BYTOM_DEBUG` environment variable to set the log level to DEBUG
